Fix: (hack) single solar inverter ignores shading
As a temporary hacky solution we assume that a single DPL controller solar inverter can always increase its power to the maximum. This is needed because a lot of people only use a single solar inverter and are having issues that the inverter gets stuck at low limits. We need to get rid of this hack at some point in time and provide a real solution for the problem.